% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node.R
\name{add_node}
\alias{add_node}
\title{Generates the SQL to add a node to the database}
\usage{
add_node(data, identifier = NA)
}
\arguments{
\item{data}{Data to be added to the node in a list format}

\item{identifier}{The identifier for the node}
}
\description{
Generates the SQL to add a node to the database
}
