% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{atomic}
\alias{atomic}
\title{An atomic transaction wrapper function}
\usage{
atomic(db_file, sql_statement)
}
\arguments{
\item{db_file}{The name of the SQLite database}

\item{sql_statement}{The SQL statement to execute}
}
\value{
Either the query results or NA for executed SQL statements
}
\description{
An atomic transaction wrapper function
}
