% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{connect_nodes}
\alias{connect_nodes}
\title{Add a node to the database}
\usage{
connect_nodes(source_id, target_id, properties = list())
}
\arguments{
\item{source_id}{Source node's id}

\item{target_id}{Target node's id}

\item{properties}{Edge properties (optional)}
}
\description{
Add a node to the database
}
