% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_inbound_neighbors.R
\name{find_inbound_neighbors}
\alias{find_inbound_neighbors}
\title{Generates the SQL to find the inbound neighbors for a node in the database}
\usage{
find_inbound_neighbors(identifier)
}
\arguments{
\item{identifier}{The identifier for the node}
}
\value{
A SQL statement to find the inbound neighbors
}
\description{
Generates the SQL to find the inbound neighbors for a node in the database
}
