% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Generates the SQL to find the neighbors for a node in the database}
\usage{
find_neighbors(identifier)
}
\arguments{
\item{identifier}{The identifier for the node}
}
\description{
Generates the SQL to find the neighbors for a node in the database
}
