% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_node.R
\name{find_node}
\alias{find_node}
\title{Generates the SQL to find a node from the database}
\usage{
find_node(identifier)
}
\arguments{
\item{identifier}{The identifier for the node}
}
\description{
Generates the SQL to find a node from the database
}
