% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{find_nodes}
\alias{find_nodes}
\title{Generate SQL to find nodes matching a criteria}
\usage{
find_nodes(data, where_fn = "search_where", search_fn = "search_equals")
}
\arguments{
\item{data}{A list of data that are the search criteria}

\item{where_fn}{The function to use in the SQL WHERE clause.  Valid values are: search_where (default) or search_like}

\item{search_fn}{The function to use in the search.  Valid values are: search_equals (default), search_starts_with, or search_contains}
}
\value{
A SQL statement to find nodes matching a criteria
}
\description{
Generate SQL to find nodes matching a criteria
}
