% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outbound_neighbors.R
\name{find_outbound_neighbors}
\alias{find_outbound_neighbors}
\title{Generates the SQL to find the outbound neighbors for a node in the database}
\usage{
find_outbound_neighbors(identifier)
}
\arguments{
\item{identifier}{The identifier for the node}
}
\value{
A SQL statement to find outbound neighbors
}
\description{
Generates the SQL to find the outbound neighbors for a node in the database
}
