% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize a new graph database}
\usage{
initialize(db_file, schema_file = "./tests/schema.sql")
}
\arguments{
\item{db_file}{The name of the SQLite database}

\item{schema_file}{The SQL schema file (optional)}
}
\description{
Initialize a new graph database
}
