% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_id.R
\name{set_id}
\alias{set_id}
\title{Sets the id attribute in JSON data}
\usage{
set_id(identifier = NA, data)
}
\arguments{
\item{identifier}{The id}

\item{data}{The JSON data}
}
\value{
JSON ecoded data
}
\description{
Sets the id attribute in JSON data
}
