% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{traverse}
\alias{traverse}
\title{Finds the path as you traverse the graph}
\usage{
traverse(db_file, src, tgt = NA, neighbors_fn = "find_neighbors")
}
\arguments{
\item{db_file}{The name of the SQLite database}

\item{src}{The id of the source node}

\item{tgt}{The id of the target node (optional)}

\item{neighbors_fn}{The neighbor function to employ.  Valid options are find_neighbors, find_inbound_neighbors or find_outbound_neighbors (optional)}
}
\description{
Finds the path as you traverse the graph
}
