% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_node.R
\name{upsert_node}
\alias{upsert_node}
\title{Generates the SQL to upsert a node in the database}
\usage{
upsert_node(identifier, data, db_file)
}
\arguments{
\item{identifier}{The identifier for the node}

\item{data}{Data to be added to the node in a list format}

\item{db_file}{The name of the SQLite database}
}
\description{
Generates the SQL to upsert a node in the database
}
