% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Generates dot files for visualization of the graph}
\usage{
visualize(
  db_file,
  dot_file,
  path = c(),
  exclude_node_keys = c(),
  hide_node_key = FALSE,
  node_kv = " ",
  exclude_edge_keys = c(),
  hide_edge_key = FALSE,
  edge_kv = " "
)
}
\arguments{
\item{db_file}{The name of the SQLite database}

\item{dot_file}{The name of the file}

\item{path}{The path to include in the visualization}

\item{exclude_node_keys}{The node keys to exclude from the visualization}

\item{hide_node_key}{Boolean if the node key is hidden}

\item{node_kv}{The node key values}

\item{exclude_edge_keys}{The key of edges to exclude}

\item{hide_edge_key}{Boolean if the edge key is hidden}

\item{edge_kv}{The edge key values}
}
\description{
Generates dot files for visualization of the graph
}
