% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connections.R
\name{get_connections}
\alias{get_connections}
\title{Generates the SQL to find the connections for a node in the database}
\usage{
get_connections(source_id, target_id)
}
\arguments{
\item{source_id}{The identifier for the source node}

\item{target_id}{The identifier for the target node}
}
\description{
Generates the SQL to find the connections for a node in the database
}
