% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_node.R
\name{remove_node}
\alias{remove_node}
\title{Generates the SQL to remove a node from the database}
\usage{
remove_node(identifier)
}
\arguments{
\item{identifier}{The identifier for the node}
}
\description{
Generates the SQL to remove a node from the database
}
