% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_node.R
\name{upsert_node}
\alias{upsert_node}
\title{Generates the SQL to upsert a node in the database}
\usage{
upsert_node(identifier, data, db_file)
}
\arguments{
\item{identifier}{The identifier for the node}

\item{data}{Data to be added to the node in a list format}

\item{db_file}{The name of the 'SQLite' database}
}
\value{
A SQL statement to upsert a node
}
\description{
Generates the SQL to upsert a node in the database
}
\examples{
\dontrun{
library(simplegraphdb)
apple <- "apple_test.sqlite"
initialize(apple)
atomic(apple, add_node(list("name" = "Apple Computer Company", "type" = c("company", "start-up"), "founded" = "April 1, 1976"), 1))
atomic(apple, add_node(list("name" = "Steve Wozniak", "type" = c("person", "engineer", "founder")), 2))
atomic(apple, add_node(list("name" = "Steve Jobs", "type" = c("person", "designer", "founder")), 3))
atomic(apple, add_node(list("name" = "Ronald Wayne", "type" = c("person", "administrator", "founder")), 4))
atomic(apple, add_node(list("name" = "Mike Markkula", "type" = c("person", "investor")), 5))
atomic(apple, connect_nodes(2, 1, list("action" = "founded")))
atomic(apple, connect_nodes(3, 1, list("action" = "founded")))
atomic(apple, connect_nodes(4, 1, list("action" = "founded")))
atomic(apple, connect_nodes(5, 1, list("action" = "invested", "equity" = 80000, "debt" = 170000)))
atomic(apple, connect_nodes(1, 4, list("action" = "divested", "amount" = 800, "date" = "April 12, 1976")))
atomic(apple, connect_nodes(2, 3))

#Upsert some data
atomic(apple, upsert_node(2, list("nickname" = "Woz"), apple))
}
}
